+++ packages/asc-ui/src/components/Select/index.ts
@@ -1,0 +1,2 @@
+export { default as SelectStyle } from './SelectStyle'
+export { default } from './Select'